requirements.txt
flask
spacy
scikit-learn
sentence-transformers
PyPDF2
python-docx
